requirements.txt
rdflib
rdfextras
pyyaml
lxml
